Fix memory limitation method on OS X (page size was ignored) and other systems (was using MBs instead of MiBs).

+++ src/libzzuf/libzzuf.c
@@ -176,7 +176,7 @@ void _zz_init(void)
 
     tmp = getenv("ZZUF_MEMORY");
     if(tmp)
-        _zz_memory = atoll(tmp);
+        _zz_memory = atoi(tmp);
 
     tmp = getenv("ZZUF_NETWORK");
     if(tmp && *tmp == '1')
